requirements.txt
dj-database-url==0.5.0
Django==2.1.5
django-widget-tweaks==1.4.3
Markdown==3.0.1
python-decouple==3.1
